added enchantment compatibility for accessories
added 5 functions;
one to tell if a accessory should be considered by X enchantment. (example: mending, protection, thorns) as well a simpler boolean function that just allow a accessory to be affect by mending,
one to tell if a accessory should do the "doPostHurtEffects",
and one to tell if a accessory should do the "doPostDamageEffects".